Added a new class to reduce the shared code between common drivers.
